requirements.txt
aiohttp==3.11.16
colorama==0.4.6
python-dotenv==1.1.0
redis==5.2.1
websockets==15.0.1
